remove manual mode switch and add time clock
- minor wording fixes
- removed manual mode and replaced it with time clock input
the time clock input is to prevent the pump from running in the middle of the
night in auto mode, but must be ignored and the pump must continued to run
as normal when the start button is pressed.

"This time switch closes its contacts when the time is set on. It is preferable
to pull low when timer is on, however I can also pull high if that is best.
I was simply going to power the Arduino on and off with the timer, but realised
it would not work in a emergency start situation."

- the start button will trigger a 'manual fill' to the limit switches of the
tanks