restructure with projects for different nodes
pump-control.ino becomes a leonardo pump-attached device with lcd display.

The next two work together, the nano is a remote rs485 connected lcd display
which is displays information sent by the master node.

display-node.ino is arduino nano with rs485 and lcd.
master-node.ino is leonardo with rs485 and no lcd.

TODO: the message recieved by the nano is partially garbage, but I can see
some of the intended characters in place.